Incluindo todas as notas da partitura
Mudanças e aplicação do novo modelo de gravação de level design. Agora
comporta notas complexas da partitura

+++ SW/Trunk/Assets/_VoaVivaldo/Scripts/Base/Nota.cs
@@ -29,7 +29,7 @@ public class NotaInfo
 	public Timbre			timbre = Timbre.SETE;
 	public Duracao  		duracao = Duracao.SEMINIMA;
 	public int				compasso;	
-	public int 				batida;
+	public float 				batida;
 	
 	public NotaInfo()
 	{
